[Bugfix] [MultiReceive] [correct cast to uint8_t of mui8_pkgRemainingInBurst]
Postreviewed by M. Wodok

+++ library/xgpl_src/IsoAgLib/comm/Part3_DataLink/impl/stream_c.cpp
@@ -163,7 +163,7 @@ Stream_c::expectBurst(uint8_t wishingPkgs)
       // Await after is CTS has timeout value of "msci32_timeOutT2=1250; // cts -> data(TP)/dpo(ETP)"
       awaitNextStep (AwaitDpo,  msci32_timeOutT2);
       // how many pkgs are missing at all? is it more than wished?
-      mui8_pkgRemainingInBurst = STL_NAMESPACE::min<uint8_t>( uint8_t(mui32_pkgTotalSize - (mui32_pkgNextToWrite - 1)), wishingPkgs);
+      mui8_pkgRemainingInBurst = uint8_t(STL_NAMESPACE::min<uint32_t>( (mui32_pkgTotalSize - (mui32_pkgNextToWrite - 1)), wishingPkgs));
       break;
 
     case StreamTP:
@@ -173,7 +173,7 @@ Stream_c::expectBurst(uint8_t wishingPkgs)
                        ? msci32_timeOutT1 /* BAM */
                        : msci32_timeOutT2 /* dest-adr. */);
       // how many pkgs are missing at all? is it more than wished?
-      mui8_pkgRemainingInBurst = STL_NAMESPACE::min<uint8_t>( uint8_t(mui32_pkgTotalSize - (mui32_pkgNextToWrite - 1)), wishingPkgs);
+      mui8_pkgRemainingInBurst = uint8_t(STL_NAMESPACE::min<uint32_t>( (mui32_pkgTotalSize - (mui32_pkgNextToWrite - 1)), wishingPkgs));
       break;
 
 #ifdef ENABLE_MULTIPACKET_VARIANT_FAST_PACKET
